make dts file to contain all types
