+++ src/api/mutations.ts
@@ -1,14 +1,6 @@
 import axios from "axios";
 import { SERVER_URI } from "constants/api-constants";
-
-export interface ITodo {
-  id: string;
-  title: string;
-  description: string;
-  colorBadge: string;
-  opened: boolean;
-  date: Date;
-}
+import { ITodo } from "types";
 
 export const addTodo = async (newTodo: ITodo): Promise<void> => {
   await axios.post(`${SERVER_URI}/todos`, newTodo);
